requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
openai>=1.0.0
google-generativeai>=0.8.0
aiohttp>=3.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.0
rich>=13.0.0
tenacity>=8.0.0
jinja2>=3.1.0
